+++ doc/sphinx/changelog.rst
@@ -3,6 +3,12 @@ FileMover release notes
 
 Release 1.x.x series
 --------------------
+
+1.0.12
+
+  - Remove hard-coded phedex URL from the code; move it into configuration; ticket #2753
+  - replace getLFNSize with phedex call instead of DBS-DD, ticket #2754
+
 This release is based on CMS COMP CVS tag V01_00_33 as starting point for CVS to SVN
 migration.